requirements.txt
rdkit-pypi
pandas
scikit-learn
joblib
pytest
